+++ src/Features/CSharp/Portable/RemoveUnnecessarySuppression/CSharpRemoveUnnecessarySuppressionCodeFixProvider.cs
@@ -14,45 +14,29 @@ using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Editing;
+using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.RemoveUnnecessarySuppression
 {
     [ExportCodeFixProvider(LanguageNames.CSharp), Shared]
-    internal class CSharpRemoveUnnecessarySuppressionCodeFixProvider : CodeFixProvider
+    internal partial class CSharpRemoveUnnecessarySuppressionCodeFixProvider : CodeFixProvider
     {
         private const string RemoveOperator = nameof(RemoveOperator);
         private const string NegateExpression = nameof(NegateExpression);
+
+        [ImportingConstructor]
+        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+        public CSharpRemoveUnnecessarySuppressionCodeFixProvider()
+        {
+        }
 
         public override ImmutableArray<string> FixableDiagnosticIds
             => ImmutableArray.Create(IDEDiagnosticIds.RemoveUnnecessarySuppressionForIsExpressionDiagnosticId);
 
         public override FixAllProvider GetFixAllProvider()
-            => new CSharpRemoveUnnecessarySuppressionFixAllProvider(this);
-
-        private class CSharpRemoveUnnecessarySuppressionFixAllProvider : DocumentBasedFixAllProvider
-        {
-            private readonly CSharpRemoveUnnecessarySuppressionCodeFixProvider _provider;
-
-            public CSharpRemoveUnnecessarySuppressionFixAllProvider(CSharpRemoveUnnecessarySuppressionCodeFixProvider provider)
-            {
-                _provider = provider;
-            }
-
-            protected override string CodeActionTitle
-                => CSharpFeaturesResources.Remove_suppression_operators;
-
-            protected override async Task<SyntaxNode?> FixAllInDocumentAsync(FixAllContext fixAllContext, Document document, ImmutableArray<Diagnostic> diagnostics)
-            {
-                var cancellationToken = fixAllContext.CancellationToken;
-                var newDoc = await FixAllAsync(
-                    document, diagnostics,
-                    fixAllContext.CodeActionEquivalenceKey == NegateExpression,
-                    cancellationToken).ConfigureAwait(false);
-                return await newDoc.GetSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-            }
-        }
+            => new CSharpRemoveUnnecessarySuppressionFixAllProvider();
 
         public override Task RegisterCodeFixesAsync(CodeFixContext context)
         {
